correção na chave estrangeira do CursoRepository

+++ src/main/java/br/com/luisEduardo/ForumHub/repository/CursoRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CursoRepository extends JpaRepository<Curso, Integer> {
+public interface CursoRepository extends JpaRepository<Curso, Long> {
 
 
     List<Curso> nome(String nome);
